Restructure of files, private methods are now at the bottom of files
