+++ net/__init__.py
@@ -1,2 +1,3 @@
 from pinet import PiNet
 from arduinonet import ArduinoNet
+from netdata import NetData
